net: initial flow API support for headers and tables
In order to manage hardware flow tables managing agents must be
able learn the hardware capabilities and resource contraints.

Because hardware can have very different capabilities and resource
support depending on the paticular device and even skew of the
device it does not seem possible to write reasonable user space
applications without this.

To support this, the following patch exposes a set of netlink
attributes and driver ndo operations which can be used to get
the hardwares capabilities. For example consider a traditional
NIC device with support for dropping packets using a 5-tuple
match compared with a high-end switch device.

The patch can be broken into three components a 'Headers'
section a 'Tables' component and driver callbacks.

The 'Headers' components is used to expose what headers and
fields can be matched in a packet. A very basic header spec
to include ethernet, and 802.1q  may look like the following,

  header_type ethernet {
	fields {
		src_mac : 6 byte
		dst_mac : 6 byte
		ethtype : 2 byte
	}
  }

  header_type vlan_tag {
	fields {
		pcp : 3 bit
		cfi : 1 bit
		vid : 12 bit
		ethertype : 16 bit
	}
  }

This can be encoded in a set user visible structures and
netlink attributes in if_flow.h see, sw_hw_field_ref() struct
and IFLA_SW_HW_HEADERS netlink attribute.

--PARSE GRAPH TBD--
How headers are interpreted is a result of the hardware
parse graph. This needs to be exported for software to
learn valid packet layouts. For example hardware may be
able to handle only a single 802.1Q tag but not Q'in'Q
tags. Or in the tunnel case only use a set of actions on
the outer tags but not the inner tags. By inspecting the
parse graph software can learn these subtle differences
between hardware devices when it needs this level of
control.
--PARSE GRAPH TBDD--

Finally this interface represents each table as a
hw_flow_table structure. These can be encoded into netlink
using the HW_FLOW_TABLE identifiers. A flow table is
constrained by its size and the subset of headers and actions
its supports.

--FlowTable Graph TBD--
Flow tables are arranged in a topoligy as described by
the xxx structure. [...]
--FlowTable Graph TBD--

TBD:
 - add support for parse graph and table graphs, for simple
   ixgbe like devices this is not very interesting. But
   for tunnel support and multiple table it is required.
 - Understand when actions are applied. Current model
   assumes actions are applied @ the table. Is this OK?
 - Error handling is completely missing sorry its a prototype
   at best at the moment.

Signed-off-by: John Fastabend <john.fastabend@intel.com>