net: implement ixgbe getlink/setlink support for hw_flow tables
This enables ixgbe to suport the flow table gets and flow
additions. Driving this code via iproute bridge tool gives
the following output.

# ./bridge/bridge flow show dev p3p2
8: p3p2
headers:
  0: ethernet {
         src_mac:48  dst_mac:48  ethertype:16
         }
  1: vlan {
         pcp:3  cfi:1  vid:1  ethertype:16
         }
  2: ipv4 {
         version:4  ihl:4  dscp:6  ecn:2  length:8
         identification:8  flags:3  fragment_offset:13  ttl:1  protocol:8
         csum:8  src_ip:32  dst_ip:32  options:*
         }
  3: tcp {
         src_port:16  dst_port:16  seq:32  ack:32  offset:4
         reserved:3  flags:9  window:8  csum:16  urgent:16
         }
  4: metadata_t {
         egress_port:32  egress_queue:32
         }

table 0:1 - l2_table:0 src 0 size 128
  matches:
         field: ethernet [dst_mac]
  actions:
         0: set_egress_port ( u32 port_id  )
  flows:
         id: 99  prio: 98
           matches:
                 ethernet.dst_mac = 00:1b:21:69:9f:09 (0000000069211b00)  <-- driver mask reporting broke
           actions:
         0: set_egress_port ( u32 port  )
table 1:1 - fdir_table:1 src 0 size 2000
  matches:
         field: vlan [vid ethertype]
         field: ipv4 [src_ip dst_ip]
         field: tcp [src_port dst_port flags]
  actions:
         0: set_egress_queue ( u32 queue  )
         1: drop_packet (  )
  flows:

# ./bridge/bridge flow add dev p3p2 \
  table l2_fwd match eth.dst 00:1b:21:69:9f:10 -1 action ingress 1

Here the show command lists the available headers the device
supports in this case, ethernet, vlan, ipv4, tcp and some metadata
to set the egress port and egress queue.

The pipeline is organized into two tables the first table supports
using the destination MAC address to forward to a port. Where a
port index is either VMDQ index or a Virtual Function (VF) associated
with SR-IOV. The index' values _should_ be reported via the standard
'ip link' command set. At the moment this may need a bit of work to
clean up the accelerated MACVLAN and SR-IOV indexing.

Tables support a set of matches which must be a subset of the
supported headers and a set of actoins. The flow listing gives
the current entries in the table.

For ixgbe the parser/pipeline is statically configured at the
moment we don't have support for creating/removing tables at
any rate. Becuase the parser is static (normal case for devices
targeted @ end hosts?) the pipline is statically defined in
the ixgbe_pipeline.h file. The flow_table API is used to generate
the netlink encoding.

Signed-off-by: John Fastabend <john.fastabend@intel.com>