net: ixgbe: add support for flow insert
this code was written a bit quickly it needs to be cleaned up!

First use a lookup table instead of switch/switch structures.
Second verifying messages will be common task across all
implementors and can be verified by the core logic.

Hey it works though so nearly got this up and running what
is still needed.

	(1) lots of code cleanup
	(2) rename all fields to drop "HW" and "hw" names
	(3) consolidate structures of _*ref objects into a
 	    single netlink attribute instead of a attribute
	    for field in structure.
	(4) implement rocker switch hooks

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>