net: flow: various message tweaks.
Add an instance indicator to support multiple instances
of the same header in the headers graph.

Remove containers for sets of tables, headers, etc. They are not
needed and I find it simplier to simply null terminate lists.

Add some #defines to ixgbe_pipeline but will fix code generator
to do a better job instead of further cleanup the pipeline by
hand.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>